Runtime module reorg
Hopefully cleaner?

I'm pretty sure that structural records would make this better;
record definitions are annoyingly tied to their hosted module,
keeping this from being as clean as I'd like.
(…the problem might not actually be connected to records,
so much as that ~foo type~ ≠ ~('a type)['a := foo]~.
This /feels/ related to structural/nominal type differences, but
isn't quite the same?)